Restructured Project, Edit Grasp Logger
